Change typedefs for vm_offset_t, vm_size_t, and vm_pindex_t, add szmin/sz/max
* Generally change these types from unsigned int to unsigned long in
  order to better catch int/long issues on both 32 and 64 bit builds.

* Add szmin() and szmax() inlines which take and return size_t.

+++ sys/cpu/i386/include/types.h
@@ -42,10 +42,10 @@
 #include <machine/stdint.h>
 #endif
 
-typedef	unsigned int	vm_offset_t;	/* address space bounded offset */
-typedef	unsigned int	vm_size_t;	/* address space bounded size */
+typedef	unsigned long	vm_offset_t;	/* address space bounded offset */
+typedef	unsigned long	vm_size_t;	/* address space bounded size */
 typedef	__int64_t	vm_ooffset_t;	/* VM object bounded offset */
-typedef	unsigned int	vm_pindex_t;	/* physical page index */
+typedef	unsigned long	vm_pindex_t;	/* physical page index */
 typedef __uint64_t	vm_poff_t;	/* physical offset */
 typedef __uint64_t	vm_paddr_t;	/* physical addr (same as vm_poff_t) */
